pyproject.toml
[project]
name = "trash-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.12.1",
    "supabase>=2.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]
